No more Raycasting, but plenty of updates...
Now, the reason I scrapped the raycasting bit was because it was becoming too big for this project. I did add the ability for the project to add the AppPages automatically to the ArrayList<AppPage> pages in engine.Window! This makes it so less code is needed to create your projects!