Easy Hardships
I've named this update "Easy Hardships" because it's both heart
wrenching, and easier on the coder. I've added a little box which allows
the player to select their preference of window type/size. I've also
added a custom debugging function (put "-debug" in the command line in
order to enable debug mode, then put the appropriate tag in for what you
want to do).