HUGE Changes!
I'm just trying to split up the bigger classes in order to make them
more readable. Now, the handlers are put inside a HandlerCollector
class. This is so if i need to add more handlers I can do so easily.
Atop that, the collector class is used to collect the classes that I'd
want the GameWindow to see.