Cleanup and Changes...
I've removed some of the things that I'd considered Game Assets worthy.
Right now, Credit does not work, so you'll have to manually change the
AppPage to mainMenu. There is also a problem still with the
StartingPanel. I am working on this, but it's hard to figure out a
decent solution.