Major Changes!
A whole lot of changes went into this update as I had to redo the entire project from scratch in order to fix a few things. Most of the earlier issues came from having the looping go through the JFrame/JPanel. Now those are separated and I updated the EntityHandler class to include z-indexing.